Refactor so performance tests are in their own folder with needed fixtures
